Move update_columns util
