+++ py-polars/polars/io/csv/batched_reader.py
@@ -9,6 +9,7 @@ from polars.datatypes import (
     N_INFER_DEFAULT,
     py_type_to_dtype,
 )
+from polars.io.csv._utils import _update_columns
 from polars.utils.various import (
     _prepare_row_count_args,
     _process_null_values,
@@ -132,8 +133,7 @@ class BatchedCsvReader:
         if batches is not None:
             if self.new_columns:
                 return [
-                    pli._update_columns(pli.wrap_df(df), self.new_columns)
-                    for df in batches
+                    _update_columns(pli.wrap_df(df), self.new_columns) for df in batches
                 ]
             else:
                 return [pli.wrap_df(df) for df in batches]
